# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/operations/{id}": {
      "delete": {
        "operationId": "cancel_operation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Operation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/pipelines": {
      "post": {
        "operationId": "start_pipeline",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartPipelineResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NodeId": {
        "type": "string",
        "format": "uuid"
      },
...
      "PageId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ReadingOrder": {
        "type": "string",
        "enum": [
          "rtl",
          "ltr",
          "custom"
        ]
      },
      "Region": {
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "properties": {
          "height": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "width": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "x": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "y": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "StartPipelineRequest": {
        "type": "object",
        "required": [
          "steps"
        ],
        "properties": {
          "allowLegacyCuda": {
            "type": "boolean",
            "description": "Experimental: allow ML CUDA on pre-Ampere GPUs."
          },
          "defaultFont": {
            "type": [
              "string",
              "null"
            ]
          },
          "pages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/PageId"
            },
            "description": "`None` → whole project, `Some(pages)` → just those pages."
          },
          "readingOrder": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReadingOrder"
              }
            ]
          },
          "region": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Region",
                "description": "Optional bounding-box hint for inpainter engines (repair-brush)."
              }
            ]
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Engine ids (`inventory::submit!` ids) to run in order."
          },
          "systemPrompt": {
            "type": [
              "string",
              "null"
            ]
          },
          "targetLanguage": {
            "type": [
              "string",
              "null"
            ]
          },
          "textNodeIds": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/NodeId"
            },
            "description": "Optional text-node ids for engines that can operate on individual blocks."
          }
        }
      },
      "StartPipelineResponse": {
        "type": "object",
        "required": [
          "operationId"
        ],
        "properties": {
          "operationId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}